--- app/src/main/java/com/bruhascended/fitapp/util/FoodHistoryItem.kt
@@ -1,3 +1,0 @@
-package com.bruhascended.fitapp.util
-
-data class FoodHistoryItem(val foodName: String)
